+++ api/src/reconciliation/dtos/env-spec.dto.ts
@@ -1,0 +1,11 @@
+export class SpecDto {
+  teamName: string
+  envName: string
+  components: SpecComponentDto[]
+}
+
+export class SpecComponentDto {
+  name: string
+  type: string
+  dependsOn: string[]
+}