+++ api/src/root-team/root-team.controller.spec.ts
@@ -1,0 +1,20 @@
+import { Test, TestingModule } from '@nestjs/testing';
+import { RootTeamController } from './root-team.controller';
+import { RootTeamService } from './root-team.service';
+
+describe('RootTeamController', () => {
+  let controller: RootTeamController;
+
+  beforeEach(async () => {
+    const module: TestingModule = await Test.createTestingModule({
+      controllers: [RootTeamController],
+      providers: [RootTeamService],
+    }).compile();
+
+    controller = module.get<RootTeamController>(RootTeamController);
+  });
+
+  it('should be defined', () => {
+    expect(controller).toBeDefined();
+  });
+});