+++ api/src/root-team/root-team.module.ts
@@ -1,0 +1,16 @@
+import { Module } from '@nestjs/common';
+import { RootTeamService } from './root-team.service';
+import { RootTeamController } from './root-team.controller';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { Team } from 'src/typeorm';
+
+@Module({
+  imports: [
+    TypeOrmModule.forFeature([
+      Team
+    ])
+  ],
+  controllers: [RootTeamController],
+  providers: [RootTeamService]
+})
+export class RootTeamModule {}