+++ api/src/team/dto/create-team.dto.ts
@@ -1,0 +1,6 @@
+import { Organization } from "src/typeorm";
+
+export class CreateTeamDto {
+  name: string;
+  organization: Organization;
+}