+++ api/src/team/dto/update-team.dto.ts
@@ -1,0 +1,6 @@
+import { PartialType } from '@nestjs/swagger';
+import { CreateTeamDto } from './create-team.dto';
+
+export class UpdateTeamDto extends PartialType(CreateTeamDto) {
+  isDeleted: boolean  
+}