+++ api/src/typeorm/environment.entity.ts
@@ -1,6 +1,8 @@
 import { Column, Entity, Index, JoinColumn, ManyToOne, OneToMany, PrimaryGeneratedColumn, UpdateDateColumn } from "typeorm";
 import { Organization } from "./Organization.entity";
 import { Component } from "./component.entity";
+import { Team } from "./team.entity";
+import { DagDto } from "src/reconciliation/dtos/environment-dag.dto";
 
 @Entity({
   name: "environment",
@@ -13,11 +15,6 @@ export class Environment {
   @Index()
   name: string;
 
-  @Column({
-    name: 'team_name',
-  })
-  teamName: string;
-
   @UpdateDateColumn({
     name: "last_reconcile_datetime",
   })
@@ -28,6 +25,20 @@ export class Environment {
 
   @OneToMany(() => Component, (component) => component.environment)
   components: Component[];
+
+  @Column({
+    type: 'json',
+    default: null
+  })
+  dag: DagDto;
+
+  @ManyToOne(() => Team, (team) => team.id, {
+    onDelete: "CASCADE"
+  })
+  @JoinColumn({
+    referencedColumnName: 'id'
+  })
+  team: Team
 
   @ManyToOne(() => Organization, (org) => org.id, {
     onDelete: "CASCADE"