+++ api/src/typeorm/index.ts
@@ -4,6 +4,7 @@ import { Organization } from './Organization.entity';
 import { ComponentReconcile } from './component-reconcile.entity';
 import { EnvironmentReconcile } from './environment-reconcile.entity';
 import { Environment } from './environment.entity';
+import { Team } from './team.entity';
 
 export const entities = [
     User,
@@ -12,9 +13,15 @@ export const entities = [
     ComponentReconcile,
     EnvironmentReconcile,
     Environment,
+    Team
 ];
 
 export {
     User,
-    Organization
+    Organization,
+    Component, 
+    ComponentReconcile,
+    EnvironmentReconcile,
+    Environment,
+    Team
 }